# docs/openapi.yaml
components:
  schemas:
    Artifact:
      additionalProperties: false
      properties:
        available:
          type: boolean
        contentType:
          type: string
        createdAt:
          format: date-time
          type: string
        epoch:
          format: int64
          type: integer
        id:
          type: string
        path:
          type: string
        sha256:
          type: string
        size:
          format: int64
          type: integer
      type: object
    ArtifactList:
      additionalProperties: false
      properties:
        artifacts:
          items:
            $ref: "#/components/schemas/Artifact"
          type:
            - array
            - "null"
      type: object
...
    ArtifactsDeleted:
      additionalProperties: false
      properties:
        deleted:
          description: "How many artifacts this request deleted: 0 when none were left."
          format: int64
          type: integer
      type: object
...
    Error:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorInfo"
      type: object
    ErrorInfo:
      additionalProperties: false
      properties:
        code:
          description: Stable error code, e.g. not_found, invalid_spec, quota_exceeded.
          type: string
        details:
          description: Every problem, when there are several (invalid_spec, secrets_required, validation).
          items:
            type: string
          type:
            - array
            - "null"
        message:
          type: string
      type: object
...
    ServerInput:
      additionalProperties: false
      properties:
        afterSync:
          description: argv run before the command when it starts after a repository sync.
          items:
            type: string
          type:
            - array
            - "null"
        command:
          description: "argv, run as the workload's user with its environment. None: only the port is exposed."
          items:
            type: string
          type:
            - array
            - "null"
        env:
          additionalProperties:
            type: string
          description: More environment for its command (not secret).
          type: object
        labels:
          additionalProperties:
            type: string
          description: "POST: free-form labels (GET /v1/servers?label=k=v)."
          type: object
        lifetime:
          description: "POST: run (default): deleted when the Run finishes for good; owner: kept, detached, until deleted (DELETE /v1/servers/{id})."
          enum:
            - run
            - owner
          type: string
        name:
          description: "Unique within the Run: 1-30 of a-z, 0-9 and -, starting with a letter, not ending in -. (POST only.)"
          type: string
        port:
          description: The port it listens on in the container.
          format: int64
          type: integer
        start:
          description: "POST: start it now (the Run must be running). Default: true when it has a command."
          type: boolean
        workdir:
          description: "Where the command runs; relative: against the workload's workdir."
          type: string
      type: object
    ServerList:
      additionalProperties: false
      properties:
        servers:
          items:
            $ref: "#/components/schemas/RunServer"
          type:
            - array
            - "null"
      type: object
...
    TenantServerList:
      additionalProperties: false
      properties:
        counts:
          additionalProperties:
            format: int64
            type: integer
          description: How many of the listed (before the state filter) are in each state.
          type: object
        servers:
          items:
            $ref: "#/components/schemas/TenantServer"
          type:
            - array
            - "null"
      type: object
...
  securitySchemes:
    apiKey:
      description: A tenant API key (luxd admin create-key), or an operator key (luxd admin create-operator-key).
      scheme: bearer
      type: http
info:
  description: |-
    The lux tenant API. Every /v1 call authenticates with an API key (`Authorization: Bearer <key>`); a key has scopes `read`, `run` and `admin`, each implying the ones before it.

    An operator key (`luxd admin create-operator-key`, scope `operator`) belongs to no tenant: it sees and acts on every tenant through the same operations. A Run's operations act in the Run's own tenant; lists span every tenant unless `?tenant=` (an id or a name) narrows them; creating needs `?tenant=`.

    Errors are JSON: `{"error": {"code": "not_found", "message": "...", "details": ["..."]}}`, with a stable `code`.
  title: lux
  version: dev
openapi: 3.1.0
paths:
...
  /v1/artifacts/{aid}:
    get:
      description: |-
        The file the Run wrote, streamed through luxd. Content-Length and X-Lux-SHA256 let a client tell a whole download from one cut short (the response is aborted, never ended cleanly, on a failure mid-way). 409 not_uploaded (with Retry-After) while it is still on its host; 410 gone once deleted (DELETE /v1/runs/{id}/artifacts).

        Needs the `read` scope.
      operationId: downloadArtifact
      parameters:
        - description: The artifact's id.
          in: path
          name: aid
          required: true
          schema:
            description: The artifact's id.
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                contentMediaType: application/octet-stream
                format: binary
                type: string
          description: The file.
          headers:
            Content-Disposition:
              description: attachment, with the file's name.
              schema:
                type: string
            Content-Length:
              description: The file's size.
              schema:
                type: integer
            X-Lux-SHA256:
              description: The file's SHA-256, hex.
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "410":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gone
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Download an artifact
      tags:
        - runs
...
  /v1/runs/{id}/artifacts:
    delete:
      description: |-
        Every artifact of a terminated Run, from storage; retention never deletes artifacts. Their listing stays, available false, and a download answers 410 gone. Works however long ago the Run ended. Idempotent: deleted is 0 once none are left. 409 not_terminal on any other Run (a stopped, lost, failed or succeeded one can still be resumed).

        Needs the `run` scope.
      operationId: deleteArtifacts
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactsDeleted"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Delete a Run's artifacts
      tags:
        - runs
    get:
      description: Needs the `read` scope.
      operationId: listArtifacts
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactList"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: List a Run's artifacts
      tags:
        - runs
...
  /v1/runs/{id}/servers:
    get:
      description: Needs the `read` scope.
      operationId: listServers
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerList"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: List a Run's servers
      tags:
        - servers
    post:
      description: |-
        A named port of the Run, with an optional command lux runs in its container. It starts now if start (default: when it has a command), which needs the Run running. Any state but finished may add one. One without a command is watched: ready whenever its port accepts connections while the Run runs.

        Needs the `run` scope.
      operationId: addServer
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunServer"
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Add a server to a Run
      tags:
        - servers
  /v1/runs/{id}/servers/{name}:
    delete:
      description: |-
        Stops it first. A server of lifetime owner is not the Run's to remove: 409 lifetime_owner (detach or delete it through /v1/servers/{id}).

        Needs the `run` scope.
      operationId: removeServer
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
        - description: The server's name.
          in: path
          name: name
          required: true
          schema:
            description: The server's name.
            type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Remove a server
      tags:
        - servers
    put:
      description: |-
        Its port, command, workdir and env. A running one keeps what it was started with until it is started again.

        Needs the `run` scope.
      operationId: putServer
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
        - description: The server's name.
          in: path
          name: name
          required: true
          schema:
            description: The server's name.
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunServer"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Change a server
      tags:
        - servers
...
  /v1/servers:
    get:
      description: |-
        Newest first, with their derived state.

        Needs the `read` scope.
      operationId: listTenantServers
      parameters:
        - description: "Operator keys: only this tenant (id or name). Tenant keys ignore it."
          explode: false
          in: query
          name: tenant
          schema:
            description: "Operator keys: only this tenant (id or name). Tenant keys ignore it."
            type: string
        - description: "Only servers in these states (comma-separated): ready, waking, asleep, stopped, unreachable, exited, no answer."
          explode: false
          in: query
          name: state
          schema:
            description: "Only servers in these states (comma-separated): ready, waking, asleep, stopped, unreachable, exited, no answer."
            type: string
        - description: Only servers with this label (key=value); repeat to require several.
          explode: true
          in: query
          name: label
          schema:
            description: Only servers with this label (key=value); repeat to require several.
            items:
              type: string
            type:
              - array
              - "null"
        - description: Only servers that wake so.
          explode: false
          in: query
          name: wake
          schema:
            description: Only servers that wake so.
            enum:
              - request
              - never
              - ""
            type: string
        - description: Only servers attached to this Run.
          explode: false
          in: query
          name: run
          schema:
            description: Only servers attached to this Run.
            type: string
        - description: "Only the server of this hostname: a full name under the preview domain or one relative DNS label."
          explode: false
          in: query
          name: hostname
          schema:
            description: "Only the server of this hostname: a full name under the preview domain or one relative DNS label."
            type: string
        - description: At most this many, newest first, 1 to 1000 (default 500).
          explode: false
          in: query
          name: limit
          schema:
            description: At most this many, newest first, 1 to 1000 (default 500).
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantServerList"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: List the tenant's servers
      tags:
        - servers
    post:
      description: |-
        A named URL that reaches a port in a Run, independent of Runs: attach it to one (runId, or POST .../attach), detach it, and it outlives them (lifetime owner). With wake request, a signed-in request while no running Run serves it emits server.wake_requested on GET /v1/events for its owner, once per wake; lux never starts a Run itself.

        Needs the `run` scope.
      operationId: createServer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantServer"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Create a server
      tags:
        - servers
...